chore: restructure sound curve loader to load generic 2D graphs instead
