--- src/ObjLoading/AssetLoading/SndCurve/SndCurveReader.h
@@ -1,32 +1,0 @@
-#pragma once
-
-#include "Utils/ClassUtils.h"
-
-#include <istream>
-#include <memory>
-#include <vector>
-
-class SndCurveReader
-{
-public:
-    class Result
-    {
-    public:
-        struct Knot
-        {
-            double m_x;
-            double m_y;
-        };
-
-        std::vector<Knot> m_knots;
-    };
-
-private:
-    std::istream& m_stream;
-    const std::string& m_filename;
-
-public:
-    SndCurveReader(std::istream& stream, const std::string& filename);
-
-    _NODISCARD std::unique_ptr<Result> Read() const;
-};
