+++ src/ObjLoading/Game/IW4/AssetLoaders/AssetLoaderSndCurve.h
@@ -8,8 +8,6 @@ namespace IW4
 {
     class AssetLoaderSndCurve final : public BasicAssetLoader<ASSET_TYPE_SOUND_CURVE, SndCurve>
     {
-        static std::string GetAssetFilename(const std::string& assetName);
-
     public:
         _NODISCARD void* CreateEmptyAsset(const std::string& assetName, MemoryManager* memory) override;
         _NODISCARD bool CanLoadFromRaw() const override;
